renamed ChannelSelect container
