+++ react/containers/ChannelSelect.jsx
